requirements.txt
# OCR Workflow Requirements
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OCR
pytesseract==0.3.10
pdf2image==1.16.0
Pillow==10.0.0

# Machine Learning
transformers==4.37.0
torch==2.1.0
sentencepiece==0.1.99

# Image Processing
opencv-python==4.8.1.0
numpy==1.24.3

# Queue & Database
redis==5.0.1
bull==3.8.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Utilities
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-bidi==1.3.0

# Frontend build tools (for reference)
# npm install vue@3 primevue @heroicons/vue

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0
black==23.12.0
